[New Features] Refactor converter (#4314)
* update converter

* complete bert-converter

* update lint

* update lint

* enable test requirements

* Update tests/transformers/bert/test_converter.py


* refactor converter

* complete bert-converter

* add Converter for compatibility

* complete bert converter

* update model-utils

* fix by comment

* update makefile

* update url-fetching

* fix test_converter

* complete bert convert

* add compatibility for converter

* update bert compatibility testing

* enable torch checkpoint

* update config

* add blank line to requirements.txt
