+++ paddlenlp/transformers/__init__.py
@@ -34,7 +34,6 @@ from .export import export_model
 from .bert.modeling import *
 from .bert.tokenizer import *
 from .bert.configuration import *
-from .bert.converter import *
 
 # isort: split
 from .gpt.modeling import *
@@ -42,10 +41,8 @@ from .gpt.tokenizer import *
 from .roberta.modeling import *
 from .roberta.tokenizer import *
 from .roberta.configuration import *
-from .roberta.converter import *
 from .electra.modeling import *
 from .electra.tokenizer import *
-from .electra.converter import *
 from .albert.modeling import *
 from .albert.tokenizer import *
 from .bit.modeling import *
@@ -144,7 +141,6 @@ from .unimo.tokenizer import *
 from .unimo.configuration import *
 from .xlnet.modeling import *
 from .xlnet.tokenizer import *
-from .xlnet.converter import *
 from .xlm.modeling import *
 from .xlm.tokenizer import *
 from .gau_alpha.modeling import *
@@ -166,7 +162,6 @@ from .clip.modeling import *
 from .clip.feature_extraction import *
 from .clip.tokenizer import *
 from .clip.procesing import *
-from .clip.converter import *
 from .clip.image_processing import *
 from .gptj.modeling import *
 from .gptj.tokenizer import *
