+++ paddlenlp/utils/env.py
@@ -57,3 +57,8 @@ SUCCESS_STATUS = 0
 
 LEGACY_CONFIG_NAME = "model_config.json"
 CONFIG_NAME = "config.json"
+PYTORCH_WEIGHT_FILE_NAME = "pytorch_model.bin"
+PADDLE_WEIGHT_FILE_NAME = "model_state.pdparams"
+
+# for conversion
+ENABLE_TORCH_CHECKPOINT = os.getenv("ENABLE_TORCH_CHECKPOINT", True)
