+++ tests/requirements.txt
@@ -1,4 +1,6 @@
 parameterized
 sentencepiece
 regex
+torch>=1.5
+transformers
 fast_tokenizer_python
