+++ tests/transformers/bert/test_modeling.py
@@ -19,6 +19,7 @@ import tempfile
 import unittest
 from typing import List
 
+import numpy as np
 import paddle
 from parameterized import parameterized_class
 
@@ -39,7 +40,7 @@ from paddlenlp.transformers.bert.configuration import BertConfig
 from paddlenlp.transformers.model_utils import PretrainedModel
 from paddlenlp.utils import install_package, uninstall_package
 
-from ...testing_utils import slow
+from ...testing_utils import require_package, slow
 from ..test_configuration_common import ConfigTester
 from ..test_modeling_common import (
     ModelTesterMixin,
@@ -585,6 +586,31 @@ class BertCompatibilityTest(unittest.TestCase):
 
         model = AutoModelForQuestionAnswering.from_pretrained("bert-base-uncased", dropout=0.3)
         self.assertEqual(model.dropout.p, 0.3)
+
+    @require_package("transformers", "torch")
+    def test_bert_converter(self):
+        with tempfile.TemporaryDirectory() as tempdir:
+
+            # 1. create commmon input
+            input_ids = np.random.randint(100, 200, [1, 20])
+
+            # 2. forward the paddle model
+            from paddlenlp.transformers import BertModel
+
+            paddle_model = BertModel.from_pretrained("hf-internal-testing/tiny-random-BertModel", cache_dir=tempdir)
+            paddle_model.eval()
+            paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
+
+            # 3. forward the torch  model
+            import torch
+            from transformers import BertModel
+
+            torch_model = BertModel.from_pretrained("hf-internal-testing/tiny-random-BertModel", cache_dir=tempdir)
+            torch_model.eval()
+            torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+            self.assertTrue(
+                np.allclose(paddle_logit.detach().cpu().numpy(), torch_logit.detach().cpu().numpy(), rtol=1e-4)
+            )
 
 
 class BertModelIntegrationTest(ModelTesterPretrainedMixin, unittest.TestCase):
